+++ packages/expo-core/android/src/main/java/expo/core/BasePackage.java
@@ -7,6 +7,7 @@ import java.util.List;
 
 import expo.core.interfaces.InternalModule;
 import expo.core.interfaces.Package;
+import expo.core.interfaces.SingletonModule;
 
 public class BasePackage implements Package {
   @Override
@@ -21,6 +22,11 @@ public class BasePackage implements Package {
 
   @Override
   public List<ViewManager> createViewManagers(Context context) {
+    return Collections.emptyList();
+  }
+
+  @Override
+  public List<SingletonModule> createSingletonModules(Context context) {
     return Collections.emptyList();
   }
 }