+++ packages/expo-core/ios/EXCore/EXModuleRegistryProvider/EXModuleRegistryProvider.h
@@ -2,12 +2,14 @@
 
 #import <Foundation/Foundation.h>
 #import <EXCore/EXModuleRegistry.h>
+#import <EXCore/EXSingletonModule.h>
 
 @interface EXModuleRegistryProvider : NSObject
 
 @property (nonatomic, weak) id<EXModuleRegistryDelegate> moduleRegistryDelegate;
 
 + (NSSet *)singletonModules;
++ (nullable EXSingletonModule *)getSingletonModuleForClass:(Class)singletonClass;
 
 - (instancetype)initWithSingletonModules:(NSSet *)modules;
 - (EXModuleRegistry *)moduleRegistryForExperienceId:(NSString *)experienceId;